requirements.txt
pandas
requests
python-dotenv
openai
